refactor(motoko): split health-check/ops-warning cluster into healthcheck.go
The stderr-attribution additions pushed motoko.go to 810 lines, tripping the
CI 800-line file-size gate. Move the self-contained HealthCheck cluster
(HealthCheck, runHealthCheck, warnIfStaleBunProcesses, clearStalePort8080,
warnIfStaleAilangLock, parseVersionOutput) verbatim into healthcheck.go:
motoko.go 810 -> 590, healthcheck.go 231. No behavior change.
